+++ annotation/CHANGELOG.md
@@ -23,8 +23,11 @@
 # 0.1.1
 - Changes in Version Compatibility
 - Annotation Location rename to prevent conflicts
-- 
+
 # 0.1.2
 - Update Versions
 - Dart 3.9+ Compatibility
 - Analyzer 8+ Compatibility
+
+# 0.1.3
+- Release **Model** and **Entity** in a seperated Annotations